INS-2588: session manager and bootstrap metrics moved
