+++ network/gateway/bootstrap/session_manager_test.go
@@ -55,9 +55,10 @@ import (
 	"testing"
 	"time"
 
-	"github.com/insolar/insolar/insolar"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
+
+	"github.com/insolar/insolar/insolar"
 )
 
 func sessionMapLen(sm SessionManager) int {
