+++ network/gateway/metrics.go
@@ -48,7 +48,7 @@
 //    whether it competes with the products or services of Insolar Technologies GmbH.
 //
 
-package bootstrap
+package gateway
 
 import (
 	"github.com/insolar/insolar/instrumentation/insmetrics"
